equality checking simplified for tests

+++ src/com/mlt/japl/fns/GteFn.java
@@ -6,7 +6,7 @@ import com.mlt.japl.iface.Array;
 public class GteFn extends PrimitiveBaseFn {
 
 	@Override
-	public int I_CC(char a, char b) {
+	public long I_CC(char a, char b) {
 		return a>=b ? 1 : 0;
 	}
